Implement the mouse listener and zone management
This commit implements the core content of this app. This tracks mouseMovement events and will warp the mouse according to the zones that are determined from the current screen positions. This includes a fallback to polling when mouseMove events are not fired (such as when mission control, launchpad, or other such apps are running).